Sjekker inntekt opp mot forventet inntekt. Dersom reell inntekt er 10 prosent eller høyere enn forventet skal det logges

+++ src/main/kotlin/no/nav/familie/ef/personhendelse/client/SakClient.kt
@@ -23,7 +23,7 @@ class SakClient(
         return response.data ?: error("Kall mot ef-sak feilet. Status=${response.status} - ${response.melding}")
     }
 
-    fun inntektForEksternId(eksternId: Int): Int {
+    fun inntektForEksternId(eksternId: Long): Int {
         val uriComponentsBuilder = UriComponentsBuilder.fromUri(uri)
             .pathSegment("api/vedtak/eksternId/$eksternId")
         val response = getForEntity<Ressurs<Int>>(uriComponentsBuilder.build().toUri())
